pyproject.toml
[tool.poetry]
name = "knmialerts"
version = "0.1.0"
description = ""
authors = ["None"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
paho-mqtt = "^2.1.0"
loguru = "^0.7.3"
python-dotenv = "^1.0.1"
xmltodict = "^0.14.2"


[tool.poetry.group.dev.dependencies]
icecream = "^2.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
